+++ src/components/RecCard1.module.css
@@ -31,7 +31,7 @@
   flex-direction: column;
   align-items: flex-start;
   justify-content: flex-start;
-  gap: var(--gap-8xs);
+  gap: var(--gap-9xs);
 }
 .div {
   position: relative;
