+++ src/pages/Homepage.module.css
@@ -137,134 +137,16 @@
   line-height: 32px;
   font-family: var(--font-roboto);
 }
+.searchContainerDiv,
+.titleDiv {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
 .titleDiv {
   align-self: stretch;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
   justify-content: flex-start;
   gap: var(--gap-7xs);
-}
-.searchFlightsB {
-  flex: 1;
-  position: relative;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  display: inline-block;
-}
-.radioFormControlLabel {
-  position: absolute;
-  top: 0;
-  left: 0;
-}
-.radioFormControlLabel1 {
-  position: absolute;
-  top: 0;
-  left: 106px;
-}
-.flightTypeDiv {
-  position: relative;
-  width: 216.53px;
-  height: 38px;
-}
-.flightOptionsDiv,
-.formTitleGroup {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-}
-.formTitleGroup {
-  overflow: hidden;
-  padding: var(--padding-9xs);
-  box-sizing: border-box;
-  gap: var(--gap-7xs);
-}
-.formTitleGroup,
-.selectOutlinedAutocomplete {
-  align-self: stretch;
-}
-.inputGroupDiv {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  padding: var(--padding-9xs);
-  box-sizing: border-box;
-  align-items: center;
-  justify-content: center;
-}
-.inputsRowDiv {
-  flex: 1;
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: flex-start;
-  gap: var(--gap-0);
-}
-.buttonDiv {
-  position: relative;
-  font-size: var(--font-size-sm);
-  letter-spacing: 0.46px;
-  line-height: 26px;
-  text-transform: uppercase;
-  font-weight: 500;
-  font-family: var(--font-roboto);
-  color: var(--color-white);
-  text-align: center;
-  display: inline-block;
-  width: 146.98px;
-}
-.buttonGroupDiv,
-.searchFlightsButton {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-.searchFlightsButton {
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  background-color: var(--color-orange-200);
-  border-radius: var(--br-3xs);
-  width: 164px;
-  height: 56px;
-  overflow: hidden;
-  transition: 0.3s;
-}
-.searchFlightsButton:hover {
-  background-color: var(--color-orange-300);
-}
-.buttonGroupDiv {
-  padding: var(--padding-9xs);
-  box-sizing: border-box;
-}
-.formInputsRow {
-  align-self: stretch;
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: flex-start;
-  gap: var(--gap-0);
-}
-.searchContainerDiv,
-.searchFormDiv {
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-}
-.searchFormDiv {
-  align-self: stretch;
-  border-radius: var(--br-sm);
-  background-color: var(--color-white);
-  padding: var(--padding-xs);
-  align-items: flex-start;
-  justify-content: flex-start;
-  gap: 4px;
-  text-align: left;
-  font-size: var(--font-size-xl);
-  color: var(--color-gray-1400);
-  font-family: var(--font-roboto);
 }
 .searchContainerDiv {
   position: absolute;
@@ -273,7 +155,7 @@
   width: 100%;
   height: 640px;
   padding: 0 140px;
-  align-items: center;
+  box-sizing: border-box;
   justify-content: center;
   gap: 43px;
 }
@@ -1061,7 +943,7 @@
   position: relative;
   height: 19px;
 }
-.buttonDiv1 {
+.buttonDiv {
   position: relative;
   font-size: var(--font-size-sm);
   letter-spacing: 0.46px;
@@ -1304,7 +1186,7 @@
   justify-content: flex-start;
   gap: var(--gap-3xs);
 }
-.buttonDiv5 {
+.buttonDiv4 {
   position: relative;
   font-size: var(--font-size-sm);
   letter-spacing: 0.46px;
@@ -1599,31 +1481,15 @@
   .hamburgerMenuIcon {
     display: flex;
   }
+  .searchContainerDiv,
   .topHeader {
     padding-left: var(--padding-lg);
     padding-right: var(--padding-lg);
     box-sizing: border-box;
   }
-  .inputsRowDiv {
-    width: 100%;
-    flex: unset;
-    align-self: stretch;
-  }
-  .searchFlightsButton {
-    margin-right: auto;
-  }
-  .buttonGroupDiv {
-    width: 100%;
-    text-align: left;
-  }
-  .formInputsRow,
-  .searchFormDiv {
-    flex-direction: column;
-  }
   .searchContainerDiv {
     padding-left: var(--padding-3xl);
     padding-right: var(--padding-3xl);
-    box-sizing: border-box;
   }
   .horizontalLineDiv {
     width: 100%;
@@ -1744,40 +1610,6 @@
   .letsExploreTravelTheWor {
     font-size: var(--font-size-8xl);
     line-height: 48px;
-  }
-  .searchFlightsB {
-    width: 100%;
-    padding-bottom: var(--padding-7xs);
-    margin-bottom: 6px;
-    border-bottom: 1px;
-    border-bottom-style: solid;
-    border-bottom-color: #eaeaea;
-    flex: unset;
-    align-self: stretch;
-  }
-  .flightTypeDiv {
-    width: 100% !important;
-  }
-  .flightOptionsDiv {
-    width: 100%;
-  }
-  .formTitleGroup {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-  .inputGroupDiv {
-    width: 100%;
-    flex: unset;
-    align-self: stretch;
-  }
-  .inputsRowDiv {
-    flex-direction: column;
-  }
-  .searchFlightsButton {
-    width: 100% !important;
-  }
-  .searchFormDiv {
-    margin-top: 20px;
   }
   .hotelButton {
     padding-bottom: var(--padding-xs);
